new leader board with back fill script

+++ src/main/java/org/cdpg/dx/aaa/leaderboard/model/AssetLeaderboardEntry.java
@@ -12,6 +12,6 @@ public record AssetLeaderboardEntry(
     String organizationId,
     String organizationName,
     String organizationType,
-    int downloads,
-    int likes,
-    int views) {}
+    long downloads,
+    long likes,
+    long views) {}
